refactor(auth): split login view helpers
- move Google legal acceptance sheet to its own file
- move debug login credentials to a debug file
- register new auth files in the Xcode project
- add shared focus clearing in LoginView
- dismiss keyboard when tapping or scrolling the login screen
- keep login, Google sign in, guest sign in, reset, and register actions focused cleanly
